+++ packages/databyss-ui/index.js
@@ -39,7 +39,7 @@ export { default as Content } from './components/Viewport/Content'
 export { default as Raw } from './components/Viewport/Raw'
 export { default as ThemedViewport } from './components/Viewport/ThemedViewport'
 export { default as Viewport } from './components/Viewport/Viewport'
-export { default as Sidebar } from './components/Sidebar/Sidebar'
+export { default as Sidebar } from './modules/Sidebar/Sidebar'
 export { default as PageContent } from './components/PageContent/PageContent'
 export {
   default as SearchContent,
